backend/requirements.txt
fastapi
uvicorn
google-adk
pydantic-settings
python-dotenv
sqlalchemy
sqlparse
pytest
httpx
PyYAML
